requirements.txt
openai 
azure-search-documents 
python-gitlab 
pandas 
sqlite3
